Trying to merge

+++ Lidgren.Network/NetStoredReliableMessage.cs
@@ -5,14 +5,14 @@ namespace Lidgren.Network
 	internal struct NetStoredReliableMessage
 	{
 		public int NumSent;
-		public double LastSent;
+		public float LastSent;
 		public NetOutgoingMessage Message;
 		public int SequenceNumber;
 
 		public void Reset()
 		{
 			NumSent = 0;
-			LastSent = 0.0;
+			LastSent = 0;
 			Message = null;
 		}
 	}
